moved components into library, created props for Navbar and Footer
