+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-    "nxConsole.nxWorkspacePath": ""
+    "nxConsole.nxWorkspacePath": "",
+    "cSpell.words": [
+        "joejohnson"
+    ]
 }
