+++ packages/components/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/components';
+export {default as Navbar} from './lib/navbar/navbar';
+export { default as Footer } from './lib/footer/footer';
