+++ packages/components/src/lib/footer/footer.tsx
@@ -1,20 +1,12 @@
 import { Link } from "react-router-dom"
+import { FooterProps } from "./footer.types"
 
-const data = [
-  {
-    name: "GitHub",
-    link: 'https://github.com/NorskJoe'
-  },
-  {
-    name: "LinkedIn",
-    link: 'https://www.linkedin.com/in/joseph-johnson-284510126/'
-  }
-]
 
-const Footer = () => {
+
+const Footer = ({links}: FooterProps) => {
   return (
     <footer>
-      {data.map((item) => (
+      {links.map((item) => (
         <Link key={item.name} to={item.link}>
           <img src={`icons/${item.name.toLowerCase()}.png`} alt={item.name} />
         </Link>
