+++ packages/components/src/lib/footer/footer.types.tsx
@@ -1,0 +1,5 @@
+import { Link } from "../shared/shared.types";
+
+export interface FooterProps {
+    links: Link[];
+}
