+++ packages/components/src/lib/navbar/navbar.types.tsx
@@ -1,0 +1,5 @@
+import { Link } from "../shared/shared.types";
+
+export interface NavbarProps {  
+    links: Link[];
+}
