+++ packages/components/src/lib/shared/shared.types.tsx
@@ -1,0 +1,4 @@
+export interface Link {
+    link: string;
+    name: string;
+}
